Move webhook code to new app
