+++ apps/backend/src/tools/database/export.ts
@@ -1,6 +1,6 @@
 import "module-alias/register";
 
-import { runApp } from "@core/server";
+import { runApp } from "@beabee/core/server";
 
 import * as models from "./anonymisers/models";
 import { anonymiseModel, clearModels } from "./anonymisers";
