+++ apps/backend/src/tools/database/import-steady.ts
@@ -10,7 +10,7 @@ import { parse } from "csv-parse";
 import { In } from "typeorm";
 
 import { getRepository } from "@beabee/core/database";
-import { runApp } from "@core/server";
+import { runApp } from "@beabee/core/server";
 import { normalizeEmailAddress } from "@beabee/core/utils/index";
 
 import ContactsService from "@beabee/core/services/ContactsService";
