+++ apps/backend/src/tools/mailchimp/sync.ts
@@ -6,7 +6,7 @@ import { Between } from "typeorm";
 
 import { getRepository } from "@beabee/core/database";
 import { log as mainLogger } from "@beabee/core/logging";
-import { runApp } from "@core/server";
+import { runApp } from "@beabee/core/server";
 
 import NewsletterService from "@beabee/core/services/NewsletterService";
 import OptionsService from "@beabee/core/services/OptionsService";
