+++ apps/backend/src/tools/test-users.ts
@@ -9,7 +9,7 @@ import moment from "moment";
 import { Brackets } from "typeorm";
 
 import { createQueryBuilder } from "@beabee/core/database";
-import { runApp } from "@core/server";
+import { runApp } from "@beabee/core/server";
 import { getActualAmount } from "@beabee/core/utils/payment";
 
 import config from "@beabee/core/config";
